# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StatChasers NFL Depth Charts API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Rank change vs the previous snapshot, if known.
 * @nullable
 */
export type DepthChartPlayerMovement = typeof DepthChartPlayerMovement[keyof typeof DepthChartPlayerMovement] | null;


export const DepthChartPlayerMovement = {
  up: 'up',
  down: 'down',
  new: 'new',
} as const;

export interface DepthChartPlayer {
  rank: number;
  name: string;
  position: string;
  /** @nullable */
  playerId?: string | null;
  /** @nullable */
  jersey?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  headshot?: string | null;
  /**
     * Link to the player's profile page on the source site, if known.
     * @nullable
     */
  profileUrl?: string | null;
  /**
     * Rank change vs the previous snapshot, if known.
     * @nullable
     */
  movement?: DepthChartPlayerMovement;
}

export interface DepthChartPositionGroup {
  position: string;
  label: string;
  players: DepthChartPlayer[];
}

export interface DepthChartGroups {
  offense: DepthChartPositionGroup[];
  defense: DepthChartPositionGroup[];
  specialTeams: DepthChartPositionGroup[];
}

export interface DepthChartTeam {
  abbr: string;
  name: string;
  /** @nullable */
  logo?: string | null;
}

/**
 * Matrix-friendly map of canonical column key -> ranked players. Keys cover every column used across the Fantasy / Offense / Defense / Special Teams tabs.
 */
export interface DepthChartPositions {[key: string]: DepthChartPlayer[]}

export interface DepthChartSnapshot {
  id: number;
  team: DepthChartTeam;
  season: number;
  source: string;
  sourceUrl: string;
  fetchedAt: string;
  groups: DepthChartGroups;
  positions: DepthChartPositions;
}

export interface RefreshResult {
  success: boolean;
  teamsRefreshed: number;
  teamsFailed: number;
  message: string;
  failedTeams?: string[];
}

export interface CacheStatus {
  totalCached: number;
  /** @nullable */
  lastRefreshAt: string | null;
  successfulTeams: string[];
  failedTeams: string[];
}