# lib/api-zod/src/generated/types/depthChartPositions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StatChasers NFL Depth Charts API
 * OpenAPI spec version: 0.1.0
 */
import type { DepthChartPlayer } from './depthChartPlayer';

/**
 * Matrix-friendly map of canonical column key -> ranked players. Keys cover every column used across the Fantasy / Offense / Defense / Special Teams tabs.
 */
export interface DepthChartPositions {[key: string]: DepthChartPlayer[]}