# lib/api-zod/src/generated/types/depthChartSnapshot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StatChasers NFL Depth Charts API
 * OpenAPI spec version: 0.1.0
 */
import type { DepthChartGroups } from './depthChartGroups';
import type { DepthChartPositions } from './depthChartPositions';
import type { DepthChartTeam } from './depthChartTeam';

export interface DepthChartSnapshot {
  id: number;
  team: DepthChartTeam;
  season: number;
  source: string;
  sourceUrl: string;
  fetchedAt: Date;
  groups: DepthChartGroups;
  positions: DepthChartPositions;
}